Improve usage message displayed when lack of args
A synopsis that has formatting, underlining each required argument as
seen in man pages, is displayed.

Though, there aren't any checks to know
if the terminal supports ANSI escape sequences, which isn't the best.
An improvement would be to use argparse, but this would probably add
too much complexity for this simple project.